ApproximatorErrors

Abstract Types

RLinearAlgebra.ApproximatorError — Type

ApproxmatorError

An abstract supertype for structures containing user-controlled parameters corresponding to methods that evaluate the quality of a Low-Rank approximation of a matrix A.

RLinearAlgebra.ApproximatorErrorRecipe — Type

ApproxmatorErrorRecipe

An abstract supertype for structures containing user-controlled parameters, linear system dependent parameters and preallocated memory corresponding to methods that evaluate the quality of a Low-Rank approximation of a matrix A.

ApproximatorError Structures

Exported Functions

RLinearAlgebra.complete_approximator_error — Function

completeapproximatorerror( error::ApproximatorError, S::CompressorRecipe, A::AbstractMatrix )

A function that produces an ApproximatorErrorRecipe from an ApproximatorError, CompressorRecipe, and AbstractMatrix.

Arguments

error::ApproximatorError, the user controlled parameters associated with the approximation error.

S::CompressorRecipe, the compressor information used for the low rank approximation.

A::AbstractMatrix, the matrix being approximated.

Outputs

The ApproximatorErrorRecipe corresponding to the ApproximatorError technique.

RLinearAlgebra.compute_approximator_error — Function

computeapproximatorerror( error::ApproximatorErrorMethod, approximator::Approximator, A::AbstractMatrix )

A function that evaluates the quality of an ApproximatorRecipe.

Arguments

error::ApproximatorErrorRecipe, the method for computing the

approximation error.

approx::ApproximatorRecipe, the low rank of the approximation of the matrix.

A::AbstractMatrix, the matrix.

Outputs

This function will return an error metric for the approximation of the matrix.