Approximators

Approximators

Abstract Types

Approximator Structures

Exported Functions

Abstract Types

RLinearAlgebra.Approximator — Type

Approximator

An abstract supertype for structures that store user-controlled parameters corresponding to technqiues that form low-rank approximations of the matrix A.

RLinearAlgebra.ApproximatorRecipe — Type

ApproximatorRecipe

An abstract supertype for structures that store user-controlled parameters, linear system dependent parameters and preallocated memory corresponding to techniques that form low-rank approximations of the matrix A.

RLinearAlgebra.ApproximatorAdjoint — Type

ApproximatorAdjoint{S<:ApproximatorRecipe} <: ApproximatorRecipe

A structure for the adjoint of a approximator recipe.

Fields

-Parent::ApproximatorRecipe, the approximator that we compute the adjoint of.

Approximator Structures

Exported Functions

RLinearAlgebra.rapproximate — Function

rapproximate(approximator::Approximator, A::AbstractMatrix)

A function that computes a Low-Rank approximation of the matrix A using the information in the provided Approximator data structure.

Arguments

approximator::Approximator, an approximation technique for computing a low-rank approximation of the matrix A.

A::AbstractMatrix, the matrix being approximated.

Outputs

An ApproximatorRecipe containing a low rank approximation of the matrix A.

RLinearAlgebra.rapproximate! — Function

rapproximate!(approximator::ApproximatorRecipe, A::AbstractMatrix)

A function that computes a Low-Rank approximation of the matrix A using the information in the provided ApproximatorRecipe data structure.

Arguments

approximator::ApproximatorRecipe, a data structure for storing the low rank approximation to the matrix A.

A::AbstractMatrix, the matrix being approximated.

Outputs

Performs an inplace update of the ApproximatorRecipe.

RLinearAlgebra.complete_approximator — Function

complete_approximator(approximator::Approximator, A::AbstractMatrix)

A function that uses information in the matrix A and user-controlled parameters in the Approximator to form an ApproximatorRecipe with appropiate memory allocations.

Arguments

approximator::Approximator, a data structure containing the user-controlled parameters relating to a particular low rank approximation.

-A::AbstractMatrix, a matrix that we wish to approximate.

Outputs

Returns an ApproximatorRecipe with memory preallocated for forming and storing the desired low rank approximation.