Compressors

Compressors

Abstract Types

Compressor Structures

Exported Functions

Internal Functions

Abstract Types

RLinearAlgebra.Compressor — Type

Compressor

An abstract supertype for structures that contain user-controlled parameters corresponding to techniques that compress a matrix.

RLinearAlgebra.CompressorRecipe — Type

CompressorRecipe

An abstract supertype for structures that contain both the user-controlled parameters in the Compressor and the memory allocations necessary for applying the compression technique to a particular linear system.

RLinearAlgebra.CompressorAdjoint — Type

CompressorAdjoint{S<:CompressorRecipe} <: CompressorRecipe

A structure for the adjoint of a compression recipe.

Fields

Parent::CompressorRecipe, the CompressorRecipe the adjoint is being applied to..

Compressor Structures

Exported Functions

RLinearAlgebra.complete_compressor — Function

complete_compressor(compressor::Compressor, A::AbstractMatrix, b::AbstractVector)

A function that uses the information in the Compressor, the matrix A, and the constrant vector b to form a CompressorRecipe that can then be multiplied with matrices and vectors.

Arguments

compress::Compressor, a compressor object.

A::AbstractMatrix, a matrix that the returned CompressorRecipe may be applied to.

b::AbstractVector, a vector that the returned CompressorRecipe may be applied to.

Outputs

A CompressorRecipe that can be applied to matrices and vectors through the use of the multiplication functions.

RLinearAlgebra.update_compressor! — Function

update_compressor!(S::CompressorRecipe, A::AbstractMatrix, b::AbstractVector)

A function that updates a CompressorRecipe with new random components, possibly based on information contained in A::AbstractMatrix and b::AbstractMatrix.

Arguments

S::CompressorRecipe, A preallocated CompressorRecipe.

A::AbstractMatrix, a matrix that could be used to update the compressor.

b::AbstractVector, a vector that could be used to update the compressor.

Outputs

Will generate an updated version of S based on the information obtained from A, b. For some compression techniques that are data oblivious this simply means generating new random entries in S.

Internal Functions

RLinearAlgebra.left_mat_mul_dimcheck — Function

left_mat_mul_dimcheck(C::AbstractMatrix, S::CompressorRecipe, A::AbstractMatrix)

Function to test the dimensions of the CompressorRecipe and matrices when applying a compression matrix to a matrix from the left.

Arguments

C::AbstractMatrix, A matrix where the output will be stored.

S::CompressorRecipe, The compression matrix information.

A::AbstractMatrix, The matrix the compressor is being applied to from the left.

Outputs

Will assert an error if one of the relevant dimensions of the three inputs is incompatible with the others.

RLinearAlgebra.right_mat_mul_dimcheck — Function

right_mat_mul_dimcheck(C::AbstractMatrix, A::AbstractMatrix), S::CompressorRecipe

Function to test the dimensions of the CompressorRecipe and matrices when applying a compression matrix to a matrix from the right.

Arguments

C::AbstractMatrix, A matrix where the output will be stored.

S::CompressorRecipe, The compression matrix information.

A::AbstractMatrix, The matrix the compressor is being applied to from the right.

Outputs

Will assert an error if one of the relevant dimensions of the three inputs is incompatible with the others.

RLinearAlgebra.vec_mul_dimcheck — Function

vec_mul_dimcheck(C::AbstractMatrix, S::CompressorRecipe, A::AbstractMatrix)

Function to test the dimensions of the CompressorRecipe and matrices when applying a compression matrix to a vector.

Arguments

x::AbstractVector, A vector where the output will be stored.

S::CompressorRecipe, The compression matrix information.

A::AbstractVector, The vector that the compressor is being applied to.

Outputs

Will assert an error if one of the relevant dimensions of the three inputs is incompatible with the others.