Linear Solver Stop Criteria

Linear Solver Stop Criteria

Abstract Types

Stopping Criteria

Abstract Types

RLinearAlgebra.LinSysStopCriterion — Type

LinSysStopCriterion

Abstract supertype for specifying stopping criteria for randomized linear solver.

Stopping Criteria

RLinearAlgebra.LSStopMaxIterations — Type

LSStopMaxIterations <: LinSysStopCriterion

A structure that specifies a maximum iteration stopping criterion. That is, once a method achieves a certain number of iterations, it stops.

Fields

max_iter::Int64, the maximum number of iterations.