Low Rank Approximation Error Methods

Low Rank Approximation Error Methods

Abstract Types

Approximate Function

Abstract Types

RLinearAlgebra.ApproxErrorMethod — Type

ApproxErrorMethod

Abstract supertype for error methods for low rank approximations to matrices.

RLinearAlgebra.RangeFinderError — Type

RangeFinderError <: ApproxErrorMethod

Abstract supertype for error methods for Random Rangefinder techniques such as the random svd decomposition random qr decomposition, and the random eigen decompositions.

Approximate Function

RLinearAlgebra.error_approximate! — Method

error_approximate!(error:T where T<: ApproxErrorMethod, LinApprox::ApproxMethod, A::AbstractMatrix)

Function that computes the error of an low rank approximation using a ApproxErrorMethod.