Low Rank Approximations

Low Rank Approximations

Abstract Types

RangeFinder Methods

Interpretlatory Decomposition Methods

Nystrom Methods

Approximate Function

Abstract Types

RLinearAlgebra.ApproxMethod — Type

ApproxMethod

Abstract supertype for low rank approximations to matrices.

RLinearAlgebra.RangeFinderMethod — Type

RangeFinderMethod <: ApproxMethod

Abstract supertype for Random Rangefinder techniques such as the random svd decomposition random qr decomposition, and the random eigen decompositions.

RLinearAlgebra.IntDecompMethod — Type

 IntDecompMethod <: ApproxMethod

Abstract supertype for Interpolatory decompositions. This includes methods like the CUR decomposition.

RLinearAlgebra.NystromMethod — Type

NystromMethod <: ApproxMethod

Abstract supertype for Nystrom techniques. These can only be applied to symmetric matrices.

RangeFinder Methods

Interpretlatory Decomposition Methods

Nystrom Methods

Approximate Function

RLinearAlgebra.rapproximate — Method

rapproximate(type::T where T<:ApproxMethod,

            A::AbstractArray
           )

A common interface for specifying different strategies for forming low rank approximations of matrices using randomized techniques. This function edits the method data structure in place but will also return the decomposition as well as an error metric representing the decompositions quality, if an error technique is specified.