SolverErrors

Abstract Types

RLinearAlgebra.SolverError — Type

SolverError

An abstract supertype for structures that contain the user-controlled parameters for techniques that evaluates the quality of solution from a linear solver.

RLinearAlgebra.SolverErrorRecipe — Type

SolverErrorRecipe

An abstract supertype for structures that contain the user-controlled parameters, linear system dependent parameters, and preallocated memory for techniques that evaluate the solution to a liear solver..

SolverError Structures

Exported Functions

RLinearAlgebra.complete_solver_error — Function

complete_solver_error(
    error::SolverError, 
    ::AbstractMatrix, 
    ::AbstractVector
)

A function that generates a SolverErrorRecipe using the user defined inputs of a SolverError and information from the matrix A and vector b.

Arguments

error::SolverError, a data structure that stores the user defined parameters relating to

an error method.

A::AbstractMatrix, the coefficient matrix.

b::AbstractVector, the constant vector.

Outputs

A Float64 representing the progress of the solver.

RLinearAlgebra.compute_solver_error — Function

compute_solver_error(
    error::ErrorRecipe, 
    solver::Solver, 
    A::AbstractMatrix, 
    b::AbstractVector
)

A function that evaluates the quality of a solution from a linear solver.

Arguments

error::SolverErrorRecipe, a data structure that stores intermediate information relating to the computation of an error metric. For instance, if computing the residual this would be a buffer vector for storing the residual.

solver::SolverRecipe, the datastructure containing all the solver information.

A::AbstractMatrix, the coefficient matrix.

b::AbstractVector, the constant vector.

Outputs

A Float64 representing the progress of the solver.