Solvers

Solvers

Abstract Types

Solver Structures

Exported Functions

Abstract Types

RLinearAlgebra.Solver — Type

Solver

An abstract supertype for structures that contain the user-controlled parameters for techniques that solve a linear system $Ax = b$.

RLinearAlgebra.SolverRecipe — Type

SolverRecipe

An abstract supertype for structures that contain the user-controlled parameters, linear system dependent parameters, and preallocated memory for techniques that solve a linear system $Ax = b$.

Solver Structures

Exported Functions

RLinearAlgebra.complete_solver — Function

complete_solver(solver::Solver, x::AbstractVector, A::AbstractMatrix, b::AbstractVector)

A function that combines the information in the Solver data structure, matrix A, vector x, and vector b to for a SolverRecipe which can be used to solve the linear system.

Arguments

solver::Solver, a solver structure containing all the user defined parameters.

x::AbstractVector, a vector that will be overwritten with the solution.

A::AbstractMatrix, the coeficent matrix for the linear system.

b::AbstractVector, the constant vector for the linear system.

Outputs

Returns a SolverRecipe.

RLinearAlgebra.rsolve — Function

rsolve(
    solver::Solver, 
    x::AbstractVector, 
    A::AbstractMatrix, 
    b::AbstractVector
)

A function that solves the linear system $Ax=b$ or the least square problem $\min_x ||Ax -b||_2^2$ using solving method specified by the solver data structure.

Arguments

solver::Solver, a solver structure containing all the user defined parameters.

x::AbstractVector, a vector that will be overwritten with the solution.

A::AbstractMatrix, the coeficent matrix for the linear system.

b::AbstractVector, the constant vector for the linear system.

Outputs

x::AbstractVector, the solution to the linear system.

solver_method, the SolverRecipe generating for applying the desired solving technique.

RLinearAlgebra.rsolve! — Function

rsolve!(
    solver::SolverRecipe, 
    x::AbstractVector, 
    A::AbstractMatrix, 
    b::AbstractVector
)

A function that solve the linear system $Ax=b$ or the least square problem $\min_x ||Ax -b||_2^2$ using solving method specified by the solver data structure.

Arguments

-solver::SolverRecipe, a structure containing all relevant parameter values and memory to solve a linear system with a speciefied technique.

x::AbstractVector, a vector that will be overwritten with the solution.

A::AbstractMatrix, the coeficent matrix for the linear system.

b::AbstractVector, the constant vector for the linear system.

Outputs

The function updates solver and x in-place.