SubSolvers

SubSolvers

Abstract Types

SubSolver Structures

Exported Functions

Abstract Types

RLinearAlgebra.SubSolver — Type

SubSolver

An abstract supertype for structures specifying solution methods for a linear system or least squares problem.

RLinearAlgebra.SubSolverRecipe — Type

SubSolverRecipe

An abstract supertype for structures with pre-allocated memory for methods that solve a linear system or least squares problem.

SubSolver Structures

RLinearAlgebra.LQSolver — Type

LQSolver <: SubSolver

A type containing information relevant to solving the linear subsystems created by the Solver routines with the LQ factorization. As there are no user controlled parameters, if the user wishes to use this method they can simply specify LQSolver().

RLinearAlgebra.LQSolverRecipe — Type

LQSolverRecipe <: SubSolverRecipe{M<:AbstractArray}

A mutable type containing informtation relevant to solving the linear subsytems created by the Solver routines with the LQ factorization.

Fields

A::M, The matrix in the linear system that will be solved with the LQ solver.

RLinearAlgebra.QRSolver — Type

QRSolver <: SubSolver

A type containing information relevant to solving the linear subsystems created by the Solver routines with the QR factorization. As there are no user controlled parameters, if the user wishes to use this method they can simply specify QRSolver().

Fields

None

RLinearAlgebra.QRSolverRecipe — Type

QRSolverRecipe{M<:AbstractArray} <: SubSolverRecipe

A mutable type containing information relevant to solving the linear subsystems created by the Solver routines with the QR factorization.

Fields

A::M, The matrix in the linear system that will be solved with the QR solver.

Exported Functions

RLinearAlgebra.complete_sub_solver — Function

complete_sub_solver(solver::SubSolver, A::AbstractArray)

A function that generates a SubSolverRecipe given the arguments.

Arguments

solver::SubSolver, a user-specified sub-solving method.

A::AbstractArray, a coefficient matrix or vector.

Returns

A SubSolverRecipe object.

RLinearAlgebra.update_sub_solver! — Function

update_sub_solver!(solver::SubSolverRecipe, A::AbstractArray)

A function that updates the SubSolver in place given arguments.

Arguments

solver::SubSolverRecipe, a fully initialized realization for a linear sub-solver.

A::AbstractArray, a coefficient matrix or vector.

Returns

Modifies the SubSolverRecipe in place given the arguments.and returns nothing.