ApproximatorErrors

ApproximatorErrors

Abstract Types

ApproximatorError Structures

Exported Functions

Abstract Types

RLinearAlgebra.ApproximatorError — Type

ApproximatorError

An abstract supertype for structures containing user-controlled parameters corresponding to methods that evaluate the quality of a low-rank approximation of a matrix A.

RLinearAlgebra.ApproximatorErrorRecipe — Type

ApproximatorErrorRecipe

An abstract supertype for structures containing user-controlled parameters, matrix dependent parameters and preallocated memory corresponding to methods that evaluate the quality of a low-rank approximation of a matrix A.

ApproximatorError Structures

Exported Functions

RLinearAlgebra.complete_approximator_error — Function

complete_approximator_error(
    error::ApproximatorError, 
    approximator::ApproximatorRecipe, 
    A::AbstractMatrix
)

A function that generates an ApproximatorErrorRecipe given the arguments.

Arguments

error::ApproximatorError, a data structure containing the user-defined parameters associated with a particular low-rank approximation error method.

approximator::ApproximatorRecipe, a fully initialized realization for a low rank approximation method for a particular matrix.

A::AbstractMatrix, a target matrix for approximation.

Outputs

An ApproximatorErrorRecipe object.

RLinearAlgebra.compute_approximator_error — Function

compute_approximator_error(
    error::ApproximatorError, 
    approximator::ApproximatorRecipe, 
    A::AbstractMatrix
)

A function that computes the approximation error of an ApproximatorRecipe for a matrix A.

Arguments

error::ApproximatorError, a data structure containing the user-defined parameters associated with a particular low-rank approximation error method.

approximator::ApproximatorRecipe, a fully initialized realization for a low rank approximation method for a particular matrix.

A::AbstractMatrix, a target matrix for approximation.

Outputs

Returns the error::Float64

RLinearAlgebra.compute_approximator_error! — Function

compute_approximator_error!(
    error::ApproximatorErrorRecipe, 
    approximator::ApproximatorRecipe, 
    A::AbstractMatrix
)

A function that computes the approximation error of an ApproximatorRecipe for a matrix A.

Arguments

error::ApproximatorErrorRecipe, a fully initialized realization for a low rank approximation error method for a particular matrix.

approximator::ApproximatorRecipe, a fully initialized realization for a low rank approximation method for a particular matrix.

A::AbstractMatrix, a target matrix for approximation.

Outputs

Returns the error::Float64