Linear Samplers

Linear Samplers

Abstract Types

Vector Row Samplers

Vector Column Samplers

Sample Function

Abstract Types

RLinearAlgebra.LinSysSampler — Type

LinSysSampler

Abstract supertype for sampling, sketching or deterministically selecting components of a linear system.

Aliases

LinSysSketch

LinSysSelect

RLinearAlgebra.LinSysVecRowSampler — Type

LinSysVecRowSampler <: LinSysSampler

Abstract supertype for sampling, sketching or deterministically selecting a row space element from a linear system.

Aliases

LinSysVecRowSketch

LinSysVecRowSelect

RLinearAlgebra.LinSysVecColSampler — Type

LinSysVecColSampler <: LinSysSampler

Abstract supertype for sampling, sketching or deterministically selecting a column space element from a linear system.

Aliases

LinSysVecColSketch

LinSysVecColSelect

RLinearAlgebra.LinSysBlkRowSampler — Type

LinSysBlkRowSampler <: LinSysSampler

Abstract supertype for sampling, sketching or deterministically selecting a collection of row space elements from a linear system.

Aliases

LinSysBlkRowSketch

LinSysBlkRowSelect

RLinearAlgebra.LinSysBlkColSampler — Type

LinSysBlkColSampler <: LinSysSampler

Abstract supertype for sampling, sketching or deterministically selecting a collection of column space elements from a linear system.

Aliases

LinSysBlkColSketch

LinSysBlkColSelect

Vector Row Samplers

RLinearAlgebra.LinSysVecRowDetermCyclic — Type

LinSysVecRowDetermCyclic <: LinSysVecRowSelect

An immutable structure without any fields. Specifies deterministic cycling through the equations of a linear system.

RLinearAlgebra.LinSysVecRowHopRandCyclic — Type

LinSysVecRowHopRandCyclic <: LinSysVecRowSelect

A mutable structure that specifies a cycling through the rows of a linear system, where the cycling order is determined randomly once the current cycling order has been used hop number of times. The solver randomly chooses the cycling order whenever necessary.

Fields

order::Union{Vector{Int64},Nothing}

hop::Int64

Constructors

LinSysVecRowHopRandCyclic() defaults to setting the order to nothing and the hop to 5 (i.e., each ordering is used five times before sampling a new ordering).

LinSysVecRowHopRandCyclic(hop::Int64) defaults to setting the order to nothing and the hop to whatever is specified by the argument.

RLinearAlgebra.LinSysVecRowOneRandCyclic — Type

LinSysVecRowOneRandCyclic <: LinSysVecRowSelect

A mutable structure with a field to store a cycling order. Randomly specifies a cycling order over the equations of a linear system. Once this ordering is specified, the ordering is kept fixed.

Fields

order::Union{Vector{Int64},Nothing}

Calling LinSysVecRowOneRandCyclic() defaults to setting order to nothing. The sample function will handle the re-initialization of the fields once the system is provided.

RLinearAlgebra.LinSysVecRowPropToNormSampler — Type

LinSysVecRowPropToNormSampler{T} <: LinSysVecRowSampler where T <: Categorical

A parametric mutable structure that specifies sampling from the rows of the equation where the probability of selecting a given equation is proportional to the sum of squares of the coefficients of the given equation. The solver will appropriately initialize the distribution.

See Strohmer, T., Vershynin, R. A Randomized Kaczmarz Algorithm with Exponential Convergence. J Fourier Anal Appl 15, 262 (2009). https://doi.org/10.1007/s00041-008-9030-4

Aliases

LinSysVecRowSVSampler

Fields

dist::T, a categorical probability distribution.

Calling LinSysVecRowPropToNormSampler() or LinSysVecRowSVSampler() defaults dist to Categorical(1.0).

RLinearAlgebra.LinSysVecRowRandCyclic — Type

LinSysVecRowRandCyclic <: LinSysVecRowSelect

A mutable structure with a field to store a cycling order. Randomly specifies a cycling order the equations of a linear system. Once this ordering is exhausted by the solver, a new random ordering is specified. This process is repeated

Fields

order::Union{Vector{Int64},Nothing}

Calling LinSysVecRowOneRandCyclic() defaults to setting order to nothing. The sample function will handle the re-initialization of the fields once the system is provided.

RLinearAlgebra.LinSysVecRowUnidSampler — Type

LinSysVecRowUnidSampler <: LinSysVecRowSampler

An immutable structure without fields that specifies randomly cycling from the rows of a linear system with uniform probability and with replacement.

Vector Column Samplers

RLinearAlgebra.LinSysVecColDetermCyclic — Type

LinSysVecColDetermCyclic <: LinSysVecColSelect

An immutable structure without any fields. Specifies deterministic cycling through the columns of a linear system.

Sample Function

RLinearAlgebra.sample — Method

sample(type::T where T<:LinSysSampler,
    A::AbstractArray,
    b::AbstractVector,
    x::AbstractVector,
    iter::Int64)

A common interface for specifying different strategies for sampling, selecting or sketching a linear system specified by A and b. The type argument is used to select the an appropriately defined strategy. The argument x is the current iterate value for the solution. The arguent iter is the iteration counter.

The value(s) returned by sample depend on the subtype of LinSysSampler being used. Specifically,

For T<:LinSysVecRowSampler, a vector in the row space of A and constant are returned

For T<:LinSysVecColSampler, a vector of length(x), the matrix A, and a scalar-valued residual are returned.