Linear Solver Logs

Linear Solver Logs

Abstract Types

Loggers

Log Function

Abstract Types

RLinearAlgebra.LinSysSolverLog — Type

LinSysSolverLog

Abstract supertype for logging information about a randomized linear system solver.

Note, all descendants that are mutable structs must have fields

iterations::Int64, which tracks the number of iterations of the solver

converged::Bool, which tracks whether the solver has converged to the solution (by some notion of convergence that is specified by the user).

Loggers

RLinearAlgebra.LSLogOracle — Type

LSLogOracle <: LinSysSolverLog

A mutable structure that stores information about a randomized linear solver's behavior. The log assumes that the true solution of the system is known and will be supplied. The goal of this log is usually for research, development, or testing.

Fields

solution::AbstractVector, a solution to the given linear system.

collection_rate::Int64, the frequency with which to record information to append to the remaining fields, starting with the initialization (i.e., iteration 0).

error_hist::Vector{Float64}, retains a vector of numbers corresponding to the error between the iterates of the solver and the solution

error_norm::Function, a function that accepts a single vector argument and returns a scalar. Used to compute the error size.

resid_hist::Vector{Float64}, retains a vector of numbers corresponding to the residual (uses the whole system to compute the residual)

resid_norm::Function, a function that accepts a single vector argument and returns a scalar. Used to compute the residual size.

iterations::Int64, the number of iterations of the solver.

converged::Bool, a flag to indicate whether the system has converged by some measure

Constructors

Calling LSLogOracle(x_star::Vector{Float64}) sets solution = x_star, collection_rate = 1, error_hist = Float64[], error_norm = norm (Euclidean norm), resid_hist = Float64[], resid_norm = norm (Euclidean norm), iterations=-1, and converged = false.

Calling LSLogOracle(x_star::Vector{Float64}, cr::Int64) sets the structure with the same parameters as for LSLogOracle(x_star) except collection_rate = cr.

RLinearAlgebra.LSLogFull — Type

LSLogFull <: LinSysSolverLog

A mutable structure that stores information about a randomized linear solver's behavior. The log assumes that the full linear system is available for processing. The goal of this log is usually for research, development or testing as it is unlikely that the entire residual vector is readily available.

Fields

collection_rate::Int64, the frequency with which to record information to append to the remaining fields, starting with the initialization (i.e., iteration 0).

resid_hist::Vector{Float64}, retains a vector of numbers corresponding to the residual (uses the whole system to compute the residual)

resid_norm::Function, a function that accepts a single vector argument and returns a scalar. Used to compute the residual size.

iterations::Int64, the number of iterations of the solver.

converged::Bool, a flag to indicate whether the system has converged by some measure

Constructors

Calling LSLogFull() sets collection_rate = 1, resid_hist = Float64[], resid_norm = norm (Euclidean norm), iterations = -1, and converged = false.

Calling LSLogFull(cr::Int64) is the same as calling LSLogFull() except collection_rate = cr.

Log Function

RLinearAlgebra.log_update! — Method

log_update!(
    log::LinSysSolverLog,
    sampler::LinSysSampler,
    x::AbstractVector,
    samp:s:Tuple,
    iter::Int64,
    A,
    b)

A common interface for specifying different strategies for updating the log with supertype LinSysSolverLog. A sampler of supertype LinSysSampler can be used to provide specific multiple implementations for the same log type. x is the iterate value. samp is the output generated by the sample function (see linears_samplers.jl). iter is the current iteration counter. A and b specify the linear system.