SolverErrors

SolverErrors

Abstract Types

SolverError Structures

Exported Functions

Abstract Types

RLinearAlgebra.SolverError — Type

SolverError

An abstract supertype for structures that track and/or evaluate the quality of a solution for a linear system or least squares.

RLinearAlgebra.SolverErrorRecipe — Type

SolverErrorRecipe

An abstract supertype for structures that contain the user-controlled parameters, linear system dependent parameters, and preallocated memory for techniques that evaluate the solution to a linear solver.

SolverError Structures

RLinearAlgebra.FullResidual — Type

FullResidual <: SolverError

A SolverError structure for computing the norm of the full residual, $b-Ax$.

Fields

None

RLinearAlgebra.FullResidualRecipe — Type

FullResidual <: SolverErrorRecipe

A SolverErrorResidual structure for computing the norm of the full residual, $b-Ax$

Fields

residual::AbstractVector, a container for the residual $b-Ax$.

RLinearAlgebra.LSGradient — Type

LSGradient <: SolverError

A SolverError structure for monitoring the residual of the normal equations of a linear system (equivalently, the gradient of the least squares problem), $-A^\top (b - Ax)$.

Fields

None

RLinearAlgebra.LSGradientRecipe — Type

LSGradientRecipe <: SolverErrorRecipe

A SolverErrorRecipe structure for storing the residual of the normal equations of a linear system (equivalently, the gradeint of the least squares problem), $-A^\intercal (b - Ax)$.

Fields

gradient::AbstractVector, $-A^\top (b - Ax)$.

RLinearAlgebra.CompressedResidual — Type

CompressedResidual <: ErrorRecipe

A structure for the compressed residual, Sb-SAx.

Fields

None

RLinearAlgebra.CompressedResidualRecipe — Type

CompressedResidualRecipe <: ErrorRecipe

A structure containing the preallocated structures for the compressed residual, Sb-SAx.

Fields

residual::AbstractVector, a container for the compressed residual, Sb-SAx.

residual_view::SubArray, a view of the residual container to handle varying compression sizes.

Exported Functions

RLinearAlgebra.complete_error — Function

complete_error(
    error::SolverError, 
    solver::Solver,
    A::AbstractMatrix, 
    b::AbstractVector
)

A function that generates a SolverErrorRecipe given the arguments.

Arguments

error::SolverError, a user-specified solver error method.

solver::Solver, a user-specified solver method.

A::AbstractMatrix, a coefficient matrix.

b::AbstractVector, a constant vector.

Returns

A SolverErrorRecipe object.

RLinearAlgebra.compute_error — Function

compute_error(
    error::SolverErrorRecipe, 
    solver::SolverRecipe, 
    x::AbstractVector,
    A::AbstractMatrix, 
    b::AbstractVector
)

A function that evaluates the error for a proposed solution vector.

Arguments

error::SolverErrorRecipe, a fully initialized realization for a solver error method for a specific linear system or least squares problem.

solver::SolverRecipe, a fully initialized realization for a solver method for a specific linear system.

x::AbstractVector, a vector for the proposed solution.

A::AbstractMatrix, a coefficient matrix.

b::AbstractVector, a constant vector.

Returns

Returns nothing