Solvers

Solvers

Abstract Types

Solver Structures

Exported Functions

Abstract Types

RLinearAlgebra.Solver — Type

Solver

An abstract supertype for structures that contain the user-controlled parameters for methods that solve linear systems and least squares problems.

RLinearAlgebra.SolverRecipe — Type

SolverRecipe

An abstract supertype specifying a solver method with pre-allocated data structures given a coefficient matrix and constant vector.

Solver Structures

Exported Functions

RLinearAlgebra.complete_solver — Function

complete_solver(solver::Solver, x::AbstractVector, A::AbstractMatrix, b::AbstractVector)

A function that generates a SolverRecipe given the arguments.

Arguments

solver::Solver, a user-specified solver method.

x::AbstractVector, a vector for the proposed solution.

A::AbstractMatrix, a coefficient matrix.

b::AbstractVector, a constant vector.

Outputs

A SolverRecipe object.

RLinearAlgebra.rsolve! — Function

rsolve!(
    solver::SolverRecipe, 
    x::AbstractVector, 
    A::AbstractMatrix, 
    b::AbstractVector
)

A function that solves a linear system given the arguments.

Arguments

solver::SolverRecipe, a fully initialized realization for a solver method for a specific linear system.

x::AbstractVector, a vector for the proposed solution.

A::AbstractMatrix, a coefficient matrix.

b::AbstractVector, a constant vector.

Outputs

Returns nothing. Updates the SolverRecipe and x in place.

rsolve!(
    solver::Solver, 
    x::AbstractVector, 
    A::AbstractMatrix, 
    b::AbstractVector
)

A function that solves a linear system given the arguments.

Arguments

solver::Solver, a user-specified solver method.

x::AbstractVector, a vector for the proposed solution.

A::AbstractMatrix, a coefficient matrix.

b::AbstractVector, a constant vector.

Outputs

x::AbstractVector, a proposed solution to a linear system or least squares problem.

A SolverRecipe object.